Fixes for 10.9

+++ Classes/NiceController.m
@@ -130,7 +130,7 @@ id swapForWindows(id each, void* context){
     return mainWindowProxy;
 }
 
--(int)runModalOpenPanel:(NSOpenPanel *)openPanel forTypes:(NSArray *)openableFileExtensions
+-(NSInteger)runModalOpenPanel:(NSOpenPanel *)openPanel forTypes:(NSArray *)openableFileExtensions
 {
     [openPanel setAllowsMultipleSelection:YES];
     [openPanel setCanChooseDirectories:YES];
